Modify worker to send info to scheduler earlier
The worker info including "Root task" did not display in the scheduler
until add of all tasks is complete. It is useful to see what the root
task of the worker is during the potentially-long scheduling phase.

We know early on what the first task is for that worker, so we can just
send the worker info to scheduler at that time.